cchlib 1.4.7   Add cchlib-sandbox project   Some minor fix
git-svn-id: https://cchlib.googlecode.com/svn/trunk@7359 bc589049-105d-1923-a3ce-6fcb24364ed1

+++ cchlib-tools/src/main/java/org/jvnet/flamingo/slider/ui/BasicFlexiSliderUI.java
@@ -411,7 +411,7 @@ public class BasicFlexiSliderUI extends FlexiSliderUI {
                 }
             }
 
-            int bumpY = ((totalContiguousWeight == 0.0) || (extraHeight < 0.0)) ? (int) (extraHeight / rangeCount)
+            int bumpY = ((totalContiguousWeight == 0.0) || (extraHeight < 0.0)) ? /*(int)*/ (extraHeight / rangeCount)
                     : 0;
             // the first control point is at the bottom
             int labelX = slider.getPreferredSize().width + ins.left + 4;
